COMMERCE-6188 adapt to use int instead of long

+++ modules/apps/commerce/commerce-account-service/src/main/java/com/liferay/commerce/account/service/impl/CommerceAccountUserRelLocalServiceImpl.java
@@ -357,8 +357,7 @@ public class CommerceAccountUserRelLocalServiceImpl
 
 	@Override
 	public int getCommerceAccountUserRelsCount(long commerceAccountId) {
-		return (int)
-			accountEntryUserRelLocalService.
+		return accountEntryUserRelLocalService.
 				getAccountEntryUserRelsCountByAccountEntryId(commerceAccountId);
 	}
